All tests green
There were 2 tests failing, a null tests for activityschedule and the main test program with a error.
I implemented a null excetpion in the constructor for a activityschedule.
The reason teh main runner gave eroros was causet here was no database linked buit spring boot searches for a datasoruce,
solved with in memory database as datasrouve purely for the tests.